Nuklear fixes

+++ Samples/03_Misc2D/Misc2D.cpp
@@ -220,7 +220,7 @@ int main()
         //statsContainer.GetComponent<TextRendererComponent>()->SetText( stats.c_str() );
 
         scene.Render();
-        System::Draw( &spriteTex, 40, 240, 100, 100, width, height );
+        System::Draw( &spriteTex, 40, 240, 100, 100, (float)width, (float)height );
         System::DrawLines( lineHandle, lineView, lineProjection );
         scene.EndFrame();
