+++ Samples/04_Misc3D/Misc3D.cpp
@@ -380,7 +380,7 @@ int main()
                 pointLights[ pointLightIndex ].AddComponent<ae3d::PointLightComponent>();
                 pointLights[ pointLightIndex ].GetComponent<ae3d::PointLightComponent>()->SetRadius( 3 );
                 pointLights[ pointLightIndex ].AddComponent<ae3d::TransformComponent>();
-                pointLights[ pointLightIndex ].GetComponent<ae3d::TransformComponent>()->SetLocalPosition( ae3d::Vec3( row * 2, -4, -100 + col * 2 ) );
+                pointLights[ pointLightIndex ].GetComponent<ae3d::TransformComponent>()->SetLocalPosition( ae3d::Vec3( (float)row * 2, -4, -100 + (float)col * 2 ) );
 
                 scene.Add( &pointLights[ pointLightIndex ] );
                 ++pointLightIndex;
